All tests pass. Fixed bugs...
I assume that because I duplicate a working Project folder (with demo inside) to make a new one, that there is sometimes a metadata issue. For example a MonoBehaviour is 'missing' on an existing prefab.

After duplicating from now on, I'll be sure to check the new and old version of the duplicate folder to be sure they work.